Fixed deleting

+++ EPlast/EPlast.BLL/Services/City/CityService.cs
@@ -251,7 +251,10 @@ namespace EPlast.BLL.Services
         {
             var city = await _repoWrapper.City.GetFirstOrDefaultAsync(c => c.ID == cityId);
 
-            await _cityBlobStorage.DeleteBlobAsync(city.Logo);
+            if (city.Logo != null)
+            {
+                await _cityBlobStorage.DeleteBlobAsync(city.Logo);
+            }
 
             _repoWrapper.City.Delete(city);
             await _repoWrapper.SaveAsync();
